rename custom prompts dir to commands
