+++ vscode/src/commands/utils/helpers.ts
