+++ vscode/src/extension.node.ts
@@ -7,8 +7,8 @@ import { GitHistory } from '@sourcegraph/cody-shared/src/chat/recipes/git-log'
 import { SourcegraphNodeCompletionsClient } from '@sourcegraph/cody-shared/src/sourcegraph-api/completions/nodeClient'
 
 import { LocalIndexedKeywordSearch } from './chat/local-code-search'
+import { CommandsController } from './commands/CommandsController'
 import { BfgRetriever } from './completions/context/retrievers/bfg/bfg-retriever'
-import { CommandsController } from './custom-prompts/CommandsController'
 import { ExtensionApi } from './extension-api'
 import { activate as activateCommon } from './extension.common'
 import { VSCODE_WEB_RECIPES } from './extension.web'
