+++ vscode/src/non-stop/FixupTypingUI.ts
@@ -1,6 +1,6 @@
 import * as vscode from 'vscode'
 
-import { EDIT_COMMAND, menu_buttons } from '../custom-prompts/utils/menu'
+import { EDIT_COMMAND, menu_buttons } from '../commands/utils/menu'
 import { getActiveEditor } from '../editor/active-editor'
 
 import { FixupTask } from './FixupTask'
